get rid of policing
Actually, the policing command, while working reasonably,  might be too
restrictive. The buffer of the router might be never full, and packet losses
tend to happen quickly.

Instead, we could rely on the shaping with the limit argument of tbf to do the
job. However, the tbf command is first and then the netem one. Thanks to the
Mininet magics, the sender can completely ignore queue limitations...

To solve this, we just keep tbf and netem, but we first introduce netem (with
very large buffer) and then tbf. With the delay, there is no more fast path and
we see again realistic results.